requirements.txt
flask
flask-cors
requests
python-dotenv
discord.py
discord-py-interactions
interactions-wait-for
pytz
torch
Pillow